delete unused files, move RequestSection to Consultations, finish section
