--- components/common/RequestsList/index.ts
@@ -1,1 +1,0 @@
-export * from './RequestsList';
