--- components/common/RequestsList/type.ts
@@ -1,7 +1,0 @@
-export type RequestsListProps = {
-  className?: string;
-  //   type: 'family' | 'individual' | 'group';
-  isFamily?: boolean;
-  isIndividual?: boolean;
-  isGroup?: boolean;
-};
