+++ components/common/TestComponent/TestComponent.tsx
@@ -24,7 +24,6 @@ import { RequestCard } from '@/components/common/RequestsCard';
 
 import dataJson from '@/data/features.json';
 import requestData from '@/data/requests.json';
-import consultationsRequests from '@/data/consultationsRequests.json';
 import communityData from '@/data/communities.json';
 
 import { FeaturesCard } from '@/components/common/FeaturesCard';
@@ -33,14 +32,10 @@ import { BurgerMenu } from '@/components/common/BurgerMenu';
 import { useState } from 'react';
 import { SliderTestComponent } from '../Slider/SliderTestComponent';
 
-import { ConsultationsRequestCard } from '@/components/ui/ConsultationsRequestCard';
-
-// import { RequestSection } from '@/sections/Consultancy/RequestSection';
 import { CommunityCard } from '@/components/common/CommunityCard';
 
 export const TestComponent = () => {
   const { featuresList } = dataJson;
-  const { familyCards, individualCards, groupCards } = consultationsRequests;
 
   const [isMenuOpen, setIsMenuOpen] = useState(false);
   const closeMenu = () => setIsMenuOpen(false);
@@ -72,23 +67,6 @@ export const TestComponent = () => {
       </button>
       <BurgerMenu isOpen={isMenuOpen} onClose={closeMenu} />
       <SliderTestComponent />
-      <div className="container flex  flex-col gap-5 pb-6 pt-6">
-        {familyCards.map(item => {
-          return <ConsultationsRequestCard key={item.id} data={item} />;
-        })}
-      </div>
-
-      <div className="container flex  flex-col gap-5 pb-6 pt-6">
-        {individualCards.map(item => {
-          return <ConsultationsRequestCard key={item.id} data={item} />;
-        })}
-      </div>
-      <div className="container flex  flex-col gap-5 pb-6 pt-6">
-        {groupCards.map(item => {
-          return <ConsultationsRequestCard key={item.id} data={item} />;
-        })}
-      </div>
-      {/* <RequestSection /> */}
     </div>
   );
 };
